+++ src/main/java/com/vetpetmon/wyrmsofnyrus/invasion/HiveCreepSpreadFurther.java
@@ -1,6 +1,7 @@
 package com.vetpetmon.wyrmsofnyrus.invasion;
 
 import com.vetpetmon.wyrmsofnyrus.block.*;
+import com.vetpetmon.wyrmsofnyrus.block.hivecreep.BlockHivecreepBase;
 import com.vetpetmon.wyrmsofnyrus.config.Debug;
 import com.vetpetmon.wyrmsofnyrus.config.Invasion;
 import com.vetpetmon.wyrmsofnyrus.synapselib.RNG;
